+++ src/terminal_state.rs
@@ -96,20 +96,20 @@ impl TerminalState {
         ret
     }
 
-    pub fn get_width(&self) -> usize {
+    pub fn get_width(&self) -> i32 {
         self.size.width
     }
 
-    pub fn set_width(&mut self, width: usize) {
+    pub fn set_width(&mut self, width: i32) {
         self.size.width = width;
         self.reset_tabs();
     }
 
-    pub fn get_height(&self) -> usize {
+    pub fn get_height(&self) -> i32 {
         self.size.height
     }
 
-    pub fn set_height(&mut self, height: usize) {
+    pub fn set_height(&mut self, height: i32) {
         self.size.height = height;
     }
 
@@ -140,7 +140,7 @@ impl TerminalState {
         let mut i = 0;
         self.tab_stops.clear();
         while i < self.get_width() {
-            self.tab_stops.push(i as i32);
+            self.tab_stops.push(i);
             i += 8;
         }
     }
@@ -153,7 +153,7 @@ impl TerminalState {
         if i < self.tab_stops.len() {
             self.tab_stops[i]
         } else {
-            self.get_width() as i32
+            self.get_width()
         }
     }
 
@@ -188,19 +188,19 @@ impl TerminalState {
         match self.origin_mode {
             crate::OriginMode::UpperLeftCorner => {
                 if buf.is_terminal_buffer {
-                    let first = buf.get_first_visible_line() as i32;
-                    let n = (first + buf.get_height() as i32 - 1).min(first.max(caret.pos.y));
+                    let first = buf.get_first_visible_line();
+                    let n = (first + buf.get_height() - 1).min(first.max(caret.pos.y));
 
                     caret.pos.y = n;
                 }
-                caret.pos.x = caret.pos.x.clamp(0, max(0, self.get_width() as i32 - 1));
+                caret.pos.x = caret.pos.x.clamp(0, max(0, self.get_width() - 1));
             }
             crate::OriginMode::WithinMargins => {
-                let first = buf.get_first_editable_line() as i32;
-                let height = buf.get_last_editable_line() as i32 - first;
+                let first = buf.get_first_editable_line();
+                let height = buf.get_last_editable_line() - first;
                 let n = caret.pos.y.clamp(first, max(first, first + height - 1));
                 caret.pos.y = n;
-                caret.pos.x = caret.pos.x.clamp(0, max(0, self.get_width() as i32 - 1));
+                caret.pos.x = caret.pos.x.clamp(0, max(0, self.get_width() - 1));
             }
         }
     }
